pavilion/AI Buddy Program/requirements.txt
anthropic>=0.39.0
click>=8.0
gspread>=6.0
google-auth>=2.0
google-auth-oauthlib>=1.0
google-api-python-client>=2.0
python-dotenv>=1.0
rich>=13.0
pyyaml>=6.0
